Encapsulate: Is this filename stdin/stdout or not?
This isn't used in KenLM, but it is in several places in Moses.  When
opening a file, the code often needs to decide whether it was meant to
be standard output, or a regular file path.  It's tempting to
encapsulate the whole call to open the file, but there are just too many
ways to do that.