Updates from Grant:
Newer version of the tuner. It has settings that seem to improve convergence speed somewhat, and the user output is clearer.

Suggested use: For a tuning corpus of thousands of lines, first tune on just a dozen or two lines, then use the resulting weights to initialize "params" for tuning the full corpus. It probably will put you in the basin of convergence for Newton's method, and you won't waste time on many bad steps. Right now doing this requires recompiling with the different initialization, but I plan to change that.